[SocialNetwork] Bring static content and configuration of OpenShift up-to-date.
Since the OpenShift set-up uses copies of the configuration files and
the static content, these have ran out-of-date compared to those in the
root directory of the workload. This commit consist of replacing these
files with the up-to-date ones and manually cherry-picking the required
changes:

 * Use the OpenShift-specific resolver in nginx.
 * Use the OpenShift-specific domain names in the Lua scripts.

Note that I did not test the changes in OpenShift; I did, however, test
them in my Kubernetes cluster, where they work well (with similar
changes to the resolver and domain names).